+++ apps/api/src/main.ts
@@ -5,7 +5,10 @@ import { configureApp } from './configure-app';
 
 async function bootstrap() {
   const environment = loadApiEnvironment();
-  const app = await NestFactory.create(AppModule, { bufferLogs: true });
+  const app = await NestFactory.create(AppModule, {
+    bufferLogs: true,
+    bodyParser: false,
+  });
   configureApp(app);
   await app.listen(environment.API_PORT, '0.0.0.0');
 }